service/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Siddiq Akbar <siddiq@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "0.115.0"
uvicorn = "0.30.6"
asyncpg = "^0.29.0"
sqlalchemy = "^2.0.35"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
python-multipart = "^0.0.11"
sqlalchemy2-stubs = "^0.0.2a38"
pylint = "^3.3.1"

[tool.poetry.group.dev.dependencies]
black = "^24.8.0"
flake8 = "^7.1.1"
isort = "^5.13.2"
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
httpx = "^0.27.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py312']
skip-string-normalization = true

[tool.isort]
profile = "black"
line_length = 88
known_third_party = ["fastapi"]
known_first_party = ["app"]
